// src/list-item/dto/create-list-item.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEnum, IsNotEmpty } from 'class-validator';
import { i18nValidationMessage } from 'nestjs-i18n';

export enum App_Name {
  CLIENT = 'CLIENT',
  ADMIN = 'ADMIN',
}

export enum ListNames {
  CLIENTS = 'CLIENTS',
  LEADS = 'LEADS',
  APPLICANTS = 'APPLICANTS',
  CONTACTS = 'CONTACTS',
  DEALS = 'DEALS',
  TRANSACTIONS = 'TRANSACTIONS',
  CREDIT = 'CREDIT',
  TASKS = 'TASKS',
  MEETINGS = 'MEETINGS',
  CALLS = 'CALLS',
  CALLBACKS = 'CALLBACKS',
  MESSAGES = 'MESSAGES',
  DELETED_USERS = 'DELETED_USERS',
  TRADING_ACCOUNTS = 'TRADING_ACCOUNTS',
  USER_KYC_DOCUMENTS = 'USER_KYC_DOCUMENTS',
  PARTNER_KYC_DOCUMENTS = 'PARTNER_KYC_DOCUMENTS',
  ROLE = 'ROLE',
  OPERATOR = 'OPERATOR',
  PARTNER = 'PARTNER',
  MT5_ACCOUNTS = 'MT5_ACCOUNTS',
  PROCESS = 'PROCESS',
  OPPORTUNITY = 'OPPORTUNITY',
  LEAD_QUESTIONS = 'LEAD_QUESTIONS',
  AGGREGATOR = 'AGGREGATOR',
  PSP = 'PSP',
  LEADS_CALL_LOGS = 'LEADS_CALL_LOGS',
  NOTES = 'NOTES',
  EMAIL_INBOX = 'EMAIL_INBOX',
  EMAIL_SENT_DRAFT = 'EMAIL_SENT_DRAFT',
  COMPANY_BANK_ACCOUNT = 'COMPANY_BANK_ACCOUNT',
  TEMPLATES = 'TEMPLATES',
  LAYOUT = 'LAYOUT',
  USER = 'USER',
  REGULATION = 'REGULATION',
  CLIENT_TICKETS = 'CLIENT_TICKETS',
  TICKETS = 'TICKETS',
  LABELS = 'LABELS',
  EMAIL_EVENTS = 'EMAIL_EVENTS',
  MASTER_TASK = 'MASTER_TASK',
  NOTIFICATION_MESSAGES = 'NOTIFICATION_MESSAGES',
  REGULATION_EVENT = 'REGULATION_EVENT',
  REGULATION_RULE = 'REGULATION_RULE',
  REGULATION_EVENT_RULE = 'REGULATION_EVENT_RULE',
  RETENTION_VOLUME_TARGET = 'RETENTION_VOLUME_TARGET',
  CLIENT_VOLUME_TARGET = 'CLIENT_VOLUME_TARGET',
  IB_PROFILE = 'IB_PROFILE',
  IB_PROFILE_CONFIG = 'IB_PROFILE_CONFIG',
  IB_COMMISSION_REPORT = 'IB_COMMISSION_REPORT',
  OPERATOR_PRODUCTIVITY = 'OPERATOR_PRODUCTIVITY',
  MT5_ACCOUNT_SUMMARY = 'MT5_ACCOUNT_SUMMARY',
  SUB_IB_LEVEL_REPORT = 'SUB_IB_LEVEL_REPORT',
  IB_COMMISSION = 'IB_COMMISSION',
  IB_COMMISSION_CLIENT_WISE_REPORT = 'IB_COMMISSION_CLIENT_WISE_REPORT',
  IB_COMMISSION_CLIENT_DEALS_REPORT = 'IB_COMMISSION_CLIENT_DEALS_REPORT',
  AUTOMATION_CONFIG = 'AUTOMATION_CONFIG',
  EVENT_REGISTRATION = 'EVENT_REGISTRATION',
  PSP_COUNTRY_PRIORITY = 'PSP_COUNTRY_PRIORITY',
  REFERRAL_PROGRAM = 'REFERRAL_PROGRAM',
  REFERRALS = 'REFERRALS',
  BONUS = 'BONUS',
}

export class CreateListItemDto {
  @ApiProperty({ enum: ListNames, required: true })
  @IsNotEmpty({ message: i18nValidationMessage('validation.NOT_EMPTY') })
  @IsEnum(ListNames)
  name: ListNames;

  @ApiProperty({ enum: App_Name, required: true })
  @IsNotEmpty({ message: i18nValidationMessage('validation.NOT_EMPTY') })
  @IsEnum(App_Name)
  appName: App_Name;
}